FUNCT added DELICT_BG.get_tags that gets user and password from the state
Most of the functionality is copied from delicatessen.js (we will remove it from
there later)

The main reason to move this to the background are:

 * We cache the tags to avoid querying the api (and waiting a few seconds) each
   time we open the popup
 * We don't want to expose get_passwd through the requests api. For that so all
   privileged functions will run in the background closure

Something that was wrong in the previous implementation is that tags was an
array of nodes instead of an array of strings. The array of nodes couldn't be
stringified by JSON (it had circular references), and thus couldn't be sent as
request result. Also, it seemd to be way heavier to handle.